Using better responsive card grid layout for note grid
This note grid, using react-grid-layout, is responsive and packs the note
preview cards in tightly, compared to using Grid container/items or manually
assembling the columns ourselves.

The change does introduce a "findDOMNode" warning that does not exist in
react-grid-layout examples. We will need to look into that at some point in the
future, but it seems pretty harmless.