Styling fixes for mention links (#100)
The cursor would disappear when moved over a mention, because it's a void
element. A hint is being added that the mention is infact focused and selected,
in the form of background color.

Also fixed a bug where using inline-block display for long heading links caused
cursor to be at the right end of the resulting box (far past end of text).

Removed unused styling under stackLink.